Support auth-code flow (#31)
* Require client secret for auth-code flow

This is needed when exchanging an auth code for an access token.

* Add handling for auth-code flow in OIDC authenticate method

* Fix typo

* Refactor auth-method handling into methods with clear case-switching

* Display auth request URL to user

* Move module constants to utilising class

The only place these constants are used, as far as I can find by grepping,
is within the OIDCAuthenticator class.
I do not think these should be generally accessible, as they are implementation-specific
details that are not intended for user modification.

* Update README to specify pre-requisite Make target

* Regen SDK without patch applied

* Manually apply tags & metrics changes

* Update tags/metrics patch from manual changes

* Add scopes in auth-code request

* Display only auth-code URL to user, not also state

* Wrap auth-code request URL in quotes to make selection simpler

Some tools, including my terminal (urxvt), recognise something wrapped in quotes
as a single, string entity.
Without quotes, such tools will select subsections on click, e.g. the host or a query key-value pair.

* Update generated Python auth code

* Update example in README with named parameters

The 'ApiClient' has numerous parameters now, with the second one _not_
being the authenticator any more.
It is better to use named parameters to avoid others being confused by this
or falling victim to parameter changes in the future.

* Use auth-code flow for de-facto example in README

* Construct URL for auth code instead of expecting user to copy full URL

It's both easier and more generic to display just the auth code
via Deploy.
Whilst 'authlib' wants a full URL to parse, we can construct this
in the SDK.

* Add retry loop for empty input for auth code

* Add spacing to identify method sections

* Fix incorrect URL construction in auth-code flow

* Update generated files

* Rename config field to reflect semantics/usage

* Update generated files

* Re-enable use of access_token in config with deprecation warning

The Configuration class doesn't have a logger defined, as it has the settings with which to generate loggers.
Instead, we can print directly to STDERR as a best-effort alternative.