Use a later version of requests package to address a CVE
Not sure why this was pinned to 2.29.0, but it's causing later versions of the seldon package to be exposed this CVE:

https://nvd.nist.gov/vuln/detail/CVE-2023-32681

I'll be using `2.0.0` in the meantime (which doesn't have the constraint), and let you know if I see any issues with the package, but I can't think of any reason why you'd need this pin.

Also applied `make python` which appears to have included some misc. changes